# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/api/socials/authorizations/{social_type}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authorization"
                ],
                "summary": "第三方登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "social_type in [weixin]",
                        "name": "social_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "微信 access_token openid 和 code，要么传 access_token openid 要么只传 code",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/authorization.Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"token\": 1}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "authorization.Authorization": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "openid": {
                    "type": "string"
                }
            }
        },
...
        "controllers.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "自定义的异常码",
                    "type": "integer"
                },
                "data": {
                    "description": "具体响应数据\n- 无数据时，默认返回一个 {}\n- 如果是列表数据建议使用 ListData 类型\n- 不需分页的列表类型时，也建议 data.list 这样响应",
                    "type": "object"
                },
                "errors": {
                    "description": "具体错误信息\n- 只用于调试的，前端不应使用该值，因为其可能会不存在",
                    "type": "object"
                },
                "message": {
                    "description": "错误描述",
                    "type": "string"
                }
            }
        },
...
    }
}